adapt delete samples and add delete sequence in API and CLI

+++ apps/backend/rest_api/data_entry/sample_job.py
@@ -7,30 +7,54 @@ from rest_api.models import Sample2Property
 from rest_api.models import Sequence
 
 
-def delete_sample(sample_list: list):
+def delete_sequences(sequence_list: list):
+    """
+    Delete sequences by name and cascade-delete all related samples,
+    alignments, properties, annotations, etc.
+    """
     data = {}
-    # TODO delete sequence or sample + all related sequences
-    samples_qs = Sample.objects.filter(name__in=sample_list)
-    sample_ids = list(samples_qs.values_list("id", flat=True))
+    sequences_qs = Sequence.objects.filter(name__in=sequence_list)
 
-    seqhash_ids = list(
-        Sequence.objects.filter(samples__in=samples_qs).values_list("id", flat=True)
-    )
-    deleted_info = samples_qs.delete()  # returns (num_deleted, dict_of_models)
-    LOGGER.info(f"Sample IDs: {sample_ids}")
-    if DEBUG:
-        print("\nSeqhash IDs", seqhash_ids)
+    if not sequences_qs.exists():
+        LOGGER.warning(f"No sequences found for deletion: {sequence_list}")
+        return {"deleted_sequences_count": 0}
 
-    LOGGER.info(f"Deleted sample: {deleted_info}")
+    deleted_info = (
+        sequences_qs.delete()
+    )  # cascade deletion, returns (num_deleted, dict_of_models)
+    LOGGER.info(f"Deleted sequence: {deleted_info}")
 
     # Filter sequences without associated samples
     clean_unused_sequences()
 
-    # TODO: Delete Annotation
+    data["deleted_sequences_count"] = deleted_info[1].get("rest_api.Sequence", 0)
+    return data
+
+
+def delete_samples(sample_list: list):
+    """
+    delte samples and all related sequences, alignments, properties
+    """
+    data = {}
+    samples_qs = Sample.objects.filter(name__in=sample_list)
+
+    if not samples_qs.exists():
+        LOGGER.warning(f"No samples found for deletion: {sample_list}")
+        return {"deleted_samples_count": 0}
+
+    sample_ids = list(samples_qs.values_list("id", flat=True))
+    sequence_ids = list(samples_qs.values_list("sequence_id", flat=True))
+    LOGGER.info(f"Deleting samples with IDs: {sample_ids}")
+    if DEBUG:
+        print("\nAssociated sequence IDs", sequence_ids)
+
+    deleted_info = samples_qs.delete()  # returns (num_deleted, dict_of_models)
+    LOGGER.info(f"Deleted samples: {deleted_info}")
+
+    # Filter sequences without associated samples
+    clean_unused_sequences()
 
     data["deleted_samples_count"] = deleted_info[1].get("rest_api.Sample", 0)
-
-    return data
 
 
 # deprecated, just for reference.
